feat(client):dto type selector
