+++ packages/amplication-client/src/ModuleDtoProperty/DtoPropertyTypesField.tsx
@@ -1,60 +1,9 @@
-import {
-  Button,
-  EnumButtonStyle,
-  EnumItemsAlign,
-  FlexItem,
-  Icon,
-} from "@amplication/ui/design-system";
+import { EnumItemsAlign, FlexItem } from "@amplication/ui/design-system";
 import { FieldArray, FieldArrayRenderProps, getIn } from "formik";
 import { get } from "lodash";
-import { useCallback, useMemo } from "react";
+import { useMemo } from "react";
 import * as models from "../models";
-import { EnumModuleDtoPropertyType } from "../models";
 import DtoPropertyTypesSelectField from "./DtoPropertyTypesSelectField";
-
-export const typeMapping: {
-  [key in models.EnumModuleDtoPropertyType]: {
-    label: string;
-  };
-} = {
-  [EnumModuleDtoPropertyType.Boolean]: {
-    label: "Boolean",
-  },
-  [EnumModuleDtoPropertyType.DateTime]: {
-    label: "DateTime",
-  },
-  [EnumModuleDtoPropertyType.Float]: {
-    label: "Float",
-  },
-  [EnumModuleDtoPropertyType.Integer]: {
-    label: "Integer",
-  },
-  [EnumModuleDtoPropertyType.Json]: {
-    label: "Json",
-  },
-  [EnumModuleDtoPropertyType.String]: {
-    label: "String",
-  },
-  [EnumModuleDtoPropertyType.Dto]: {
-    label: "Dto",
-  },
-  [EnumModuleDtoPropertyType.Enum]: {
-    label: "Enum",
-  },
-  [EnumModuleDtoPropertyType.Null]: {
-    label: "Null",
-  },
-  [EnumModuleDtoPropertyType.Undefined]: {
-    label: "Undefined",
-  },
-};
-
-export const TYPE_OPTIONS = Object.values(EnumModuleDtoPropertyType).map(
-  (value) => ({
-    label: typeMapping[value].label,
-    value,
-  })
-);
 
 type Props = {
   name: string;
@@ -84,26 +33,16 @@ const PropertyTypes = ({
   return (
     <div>
       {errors && <div>{errors}</div>}
-      {propertyTypes?.length
-        ? propertyTypes.map((type: models.PropertyTypeDef, index: number) => (
-            <PropertyType
-              key={index}
-              index={index}
-              onRemove={remove}
-              name={name}
-              value={type}
-            />
-          ))
-        : ""}
-      <Button
-        onClick={() =>
-          push({ type: EnumModuleDtoPropertyType.String, isArray: false })
-        }
-        buttonStyle={EnumButtonStyle.Text}
-      >
-        <Icon icon="plus" />
-        Add union type
-      </Button>
+      {propertyTypes?.length &&
+        propertyTypes.map((type: models.PropertyTypeDef, index: number) => (
+          <PropertyType
+            key={index}
+            index={index}
+            onRemove={remove}
+            name={name}
+            value={type}
+          />
+        ))}
     </div>
   );
 };
@@ -116,19 +55,20 @@ type PropertyTypeProps = {
 };
 
 const PropertyType = ({ name, index, onRemove, value }: PropertyTypeProps) => {
-  const handleRemoveOption = useCallback(() => {
-    onRemove(index);
-  }, [onRemove, index]);
+  //@todo: return remove option when supporting union types
+  // const handleRemoveOption = useCallback(() => {
+  //   onRemove(index);
+  // }, [onRemove, index]);
 
   return (
     <FlexItem itemsAlign={EnumItemsAlign.Center}>
       <DtoPropertyTypesSelectField name={`${name}.${index}`} label="Type" />
-      <Button
+      {/* <Button
         type="button"
         buttonStyle={EnumButtonStyle.Text}
         icon="trash_2"
         onClick={handleRemoveOption}
-      />
+      /> */}
     </FlexItem>
   );
 };
