--- packages/amplication-client/src/ModuleDtoProperty/DtoPropertyTypesSelectField2.scss
@@ -1,40 +1,0 @@
-@import "../style/index.scss";
-
-.dto-property-types-select-field2 {
-  &__modal-wrapper {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-
-    .SelectMenu-icon {
-      margin-right: 0 !important;
-      margin-left: var(--default-spacing);
-    }
-  }
-
-  &__header {
-    padding: var(--default-spacing);
-  }
-
-  .container {
-    border-top: 1px solid var(--border-color);
-    flex: 1;
-    overflow: auto;
-    .group {
-      background-color: var(--gray-90);
-      margin-left: var(--icon-spacing);
-
-      .group-name {
-        display: flex;
-        flex-direction: row;
-        gap: var(--default-spacing);
-        padding: var(--icon-spacing);
-        // background-color: var(--gray-80);
-      }
-
-      .sub-group-name {
-        // background-color: var(--gray-70);
-      }
-    }
-  }
-}
